Move filtering functionality into separate module
